refactor `poll_until` to only ever call wasm inside a fiber (#13)
* refactor `poll_until` to only ever call wasm inside a fiber

With `wasmtime`'s `async` support enabled, all guest export calls need to happen
in a fiber because the guest may call back to the host, which may try to suspend
the fiber and return control to the top-level async executor.  In fact, even if
the guest doesn't call back to the host, epoch interruption might be enabled, in
which case the code might be instrumented with yield calls that do the same
thing.

In the process of this refactor, I discovered a couple of other issues (e.g. not
removing tasks from the "yielding" set when disposing of them) and fixed those
as well.

Fixes #9.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* disable `cargo-vet` tasks in `wasip3-prototyping.yml`

We'll address `cargo-vet` auditing when upstreaming these changes, but it's just
a distraction while we're working in this repo.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

---------

Signed-off-by: Joel Dice <joel.dice@fermyon.com>